Separate serialization from extraction code
